feat(web): optimize architecture, design system and component standards
- Decomposed massive App.tsx into focused custom hooks (useAppAuth, useAppAgents, etc.)
- Migrated hardcoded Tailwind colors to semantic tokens in tailwind.config.cjs
- Standardized UI components using Mantine Box and UnstyledButton
- Fixed linter errors and optimized hook dependencies
- Improved build stability by ignoring build artifacts in ESLint
