+++ web/src/app_utils.ts
@@ -1,0 +1,138 @@
+import { parseApiErrorMessage } from "./api";
+
+export function formatWorktreeError(err: unknown): string | null {
+  const msg = parseApiErrorMessage(err);
+  if (!msg) return null;
+  const lower = msg.toLowerCase();
+  if (!lower.includes("worktree") && !lower.includes("workdir")) return null;
+  if (lower.includes("workdir not allowed")) {
+    return "Workdir not allowed. Add the path to Safe Paths before starting the agent.";
+  }
+  if (lower.includes("worktree_repo required")) {
+    return "Worktree repo is required for the selected mode.";
+  }
+  if (lower.includes("worktree does not exist")) {
+    return "Worktree does not exist. Use Create Worktree or choose an existing workdir.";
+  }
+  if (lower.includes("workdir is not empty")) {
+    return "Workdir is not empty. Choose an empty directory for Create Worktree.";
+  }
+  if (lower.includes("git worktree add failed")) {
+    return `Git worktree add failed. ${msg}`;
+  }
+  return msg;
+}
+
+export function parseSendInputSessionMismatch(
+  message: string
+): { expected: string; running: string } | null {
+  const match = message.match(
+    /agent session mismatch:\s*expected=([^\s]+)\s+running=([^\s]+)/
+  );
+  if (!match) return null;
+  const expected = match[1]?.trim();
+  const running = match[2]?.trim();
+  if (!expected || !running) return null;
+  return { expected, running };
+}
+
+export function createAnsiRenderer(): (input: string) => string {
+  const escapeAnsiHtml = (input: string): string =>
+    input
+      .replace(/&/g, "&amp;")
+      .replace(/</g, "&lt;")
+      .replace(/>/g, "&gt;")
+      .replace(/"/g, "&quot;")
+      .replace(/'/g, "&#39;");
+  const colors: Record<number, string> = {
+    30: "#1e1e1e",
+    31: "#e06c75",
+    32: "#98c379",
+    33: "#e5c07b",
+    34: "#61afef",
+    35: "#c678dd",
+    36: "#56b6c2",
+    37: "#dcdfe4",
+    90: "#7f848e",
+    91: "#ff6b6b",
+    92: "#b2f2bb",
+    93: "#ffe066",
+    94: "#74c0fc",
+    95: "#e599f7",
+    96: "#66d9e8",
+    97: "#ffffff",
+  };
+  const bgColors: Record<number, string> = {
+    40: "#1e1e1e",
+    41: "#5c2a2a",
+    42: "#2a4a2a",
+    43: "#5c4a2a",
+    44: "#2a3a5c",
+    45: "#4a2a5c",
+    46: "#2a5c5c",
+    47: "#dcdfe4",
+    100: "#3b3b3b",
+    101: "#7a2e2e",
+    102: "#2e6a2e",
+    103: "#7a6a2e",
+    104: "#2e4a7a",
+    105: "#6a2e7a",
+    106: "#2e7a7a",
+    107: "#ffffff",
+  };
+
+  return (input: string) => {
+    const esc = "\u001b[";
+    // eslint-disable-next-line no-control-regex
+    const regex = /\u001b\[[0-9;]*m/g;
+    let lastIndex = 0;
+    let fg: string | null = null;
+    let bg: string | null = null;
+    let out = "";
+
+    const pushText = (text: string) => {
+      const safe = escapeAnsiHtml(text);
+      if (!fg && !bg) {
+        out += safe;
+        return;
+      }
+      const style = [
+        fg ? `color:${fg}` : "",
+        bg ? `background:${bg}` : "",
+      ]
+        .filter(Boolean)
+        .join(";");
+      out += `<span style="${style}">${safe}</span>`;
+    };
+
+    let match: RegExpExecArray | null;
+    while ((match = regex.exec(input)) !== null) {
+      const idx = match.index;
+      if (idx > lastIndex) {
+        pushText(input.slice(lastIndex, idx));
+      }
+      const seq = match[0].slice(esc.length, -1);
+      const parts = seq.split(";").filter(Boolean).map(Number);
+      if (parts.length === 0) {
+        fg = null;
+        bg = null;
+      } else {
+        for (const code of parts) {
+          if (code === 0) {
+            fg = null;
+            bg = null;
+          } else if (colors[code]) {
+            fg = colors[code];
+          } else if (bgColors[code]) {
+            bg = bgColors[code];
+          }
+        }
+      }
+      lastIndex = regex.lastIndex;
+    }
+    if (lastIndex < input.length) {
+      pushText(input.slice(lastIndex));
+    }
+    return out;
+  };
+}
